change types
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@10502 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/languages/mod_managed/freeswitch_wrap.cxx
@@ -26839,10 +26839,10 @@ SWIGEXPORT void SWIGSTDCALL CSharp_console_clean_log(char * jarg1) {
 }
 
 
-SWIGEXPORT void SWIGSTDCALL CSharp_msleep(unsigned long jarg1) {
-  uint32_t arg1 ;
+SWIGEXPORT void SWIGSTDCALL CSharp_msleep(unsigned int jarg1) {
+  unsigned int arg1 ;
   
-  arg1 = (uint32_t)jarg1; 
+  arg1 = (unsigned int)jarg1; 
   msleep(arg1);
 }
